Move debug demangle_helper etc into util::debug namespace/dir
