+++ hpx/runtime/serialization/detail/polymorphic_intrusive_factory.hpp
@@ -10,7 +10,7 @@
 
 #include <hpx/config.hpp>
 #include <hpx/runtime/serialization/serialization_fwd.hpp>
-#include <hpx/util/demangle_helper.hpp>
+#include <hpx/util/debug/demangle_helper.hpp>
 #include <hpx/util/detail/pp/stringize.hpp>
 #include <hpx/util/jenkins_hash.hpp>
 
@@ -157,12 +157,12 @@ namespace hpx { namespace serialization { namespace detail
 
 #define HPX_SERIALIZATION_POLYMORPHIC_TEMPLATE(Class)                         \
   HPX_SERIALIZATION_POLYMORPHIC_WITH_NAME(                                    \
-      Class, hpx::debug::type_id<Class>::typeid_.type_id();)                   \
+      Class, hpx::util::debug::type_id<Class>::typeid_.type_id();)            \
 /**/
 
 #define HPX_SERIALIZATION_POLYMORPHIC_TEMPLATE_SPLITTED(Class)                \
   HPX_SERIALIZATION_POLYMORPHIC_WITH_NAME_SPLITTED(                           \
-      Class, hpx::debug::type_id<T>::typeid_.type_id();)                       \
+      Class, hpx::util::debug::type_id<T>::typeid_.type_id();)                \
 /**/
 
 #endif
