+++ hpx/runtime/serialization/detail/polymorphic_nonintrusive_factory.hpp
@@ -16,7 +16,7 @@
 #include <hpx/traits/needs_automatic_registration.hpp>
 #include <hpx/traits/polymorphic_traits.hpp>
 #include <hpx/util/assert.hpp>
-#include <hpx/util/demangle_helper.hpp>
+#include <hpx/util/debug/demangle_helper.hpp>
 #include <hpx/util/detail/pp/stringize.hpp>
 #include <hpx/util/detail/pp/strip_parens.hpp>
 #include <hpx/util/jenkins_hash.hpp>
@@ -48,7 +48,7 @@ namespace hpx { namespace serialization { namespace detail
             static_assert(
                 traits::needs_automatic_registration<T>::value,
                 "HPX_REGISTER_ACTION_DECLARATION missing");
-            return debug::type_id<T>::typeid_.type_id();
+            return util::debug::type_id<T>::typeid_.type_id();
         }
     };
 #endif
@@ -281,7 +281,7 @@ namespace hpx { namespace serialization { namespace detail
 #define HPX_SERIALIZATION_REGISTER_CLASS_TEMPLATE(Parameters, Template)       \
     HPX_SERIALIZATION_REGISTER_CLASS_NAME_TEMPLATE(                           \
         Parameters, Template,                                                 \
-        hpx::debug::type_id<HPX_PP_STRIP_PARENS(Template) >::typeid_.type_id())\
+        hpx::util::debug::type_id<HPX_PP_STRIP_PARENS(Template) >::typeid_.type_id())\
     HPX_PP_STRIP_PARENS(Parameters) hpx::serialization::detail::register_class< \
         HPX_PP_STRIP_PARENS(Template)>                                        \
         HPX_PP_STRIP_PARENS(Template)::hpx_register_class_instance;           \
