+++ hpx/util/detail/function_registration.hpp
@@ -9,7 +9,7 @@
 #define HPX_UTIL_DETAIL_FUNCTION_REGISTRATION_HPP
 
 #include <hpx/config.hpp>
-#include <hpx/util/demangle_helper.hpp>
+#include <hpx/util/debug/demangle_helper.hpp>
 #include <hpx/util/detail/pp/stringize.hpp>
 #include <hpx/util/detail/pp/strip_parens.hpp>
